# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nuestra Medicina Personal API
  version: 0.1.0
  description: API REST del monolito para la tienda personal de eBooks.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
tags:
  - name: Health
  - name: Authentication
  - name: Books
  - name: Orders
  - name: Library
  - name: Webhooks
  - name: Admin books
paths:
  /health/live:
    get:
      tags: [Health]
      operationId: getLiveness
      responses:
        '200':
          description: El proceso HTTP está vivo.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Health'}
  /health/ready:
    get:
      tags: [Health]
      operationId: getReadiness
      responses:
        '200':
          description: La API puede acceder a PostgreSQL.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Health'}
        '503': {$ref: '#/components/responses/DatabaseUnavailable'}
  /api/v1/auth/google:
    get:
      tags: [Authentication]
      operationId: startGoogleAuthentication
      description: Inicia Authorization Code con PKCE, state y nonce.
      responses:
        '302': {description: Redirección al endpoint de autorización de Google.}
        '503':
          description: Las credenciales Google no están configuradas.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/auth/google/callback:
    get:
      tags: [Authentication]
      operationId: completeGoogleAuthentication
      parameters:
        - {name: code, in: query, schema: {type: string}}
        - {name: state, in: query, schema: {type: string}}
        - {name: error, in: query, schema: {type: string}}
      responses:
        '302': {description: Sesión creada y redirección a la biblioteca, o regreso al login si el usuario canceló.}
        '400':
          description: State, nonce, PKCE o callback inválido/expirado.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '502':
          description: No fue posible validar la identidad con Google.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/Origin'
      responses:
        '204': {description: Sesión revocada y cookie eliminada.}
        '403':
          description: Origin no permitido.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Usuario de la sesión vigente.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/books:
    get:
      tags: [Books]
      operationId: listPublishedBooks
      description: Devuelve exclusivamente libros publicados.
      responses:
        '200':
          description: Catálogo público.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BookList'}
  /api/v1/books/{slug}:
    parameters:
      - $ref: '#/components/parameters/BookSlug'
    get:
      tags: [Books]
      operationId: getPublishedBook
      description: Un borrador o libro archivado se responde como inexistente.
      responses:
        '200':
          description: Libro publicado.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
        '404': {$ref: '#/components/responses/BookNotFound'}
  /api/v1/me/books:
    get:
      tags: [Library]
      operationId: listPurchasedBooks
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Libros asociados a órdenes pagadas del usuario.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LibraryBookList'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
  /api/v1/books/{id}/download:
    get:
      tags: [Library]
      operationId: downloadPurchasedBook
      description: Autoriza la compra y delega el archivo a una ubicación interna de Nginx.
      security: [{sessionCookie: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Descarga protegida resuelta internamente por Nginx.
          headers:
            Content-Disposition: {schema: {type: string}}
            X-Accel-Redirect: {schema: {type: string}}
          content:
            application/pdf: {}
            application/epub+zip: {}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '404':
          description: El usuario no compró el libro o el archivo no está disponible.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/orders:
    post:
      tags: [Orders]
      operationId: createOrder
      description: Crea una orden con precio histórico y una preferencia de Checkout Pro.
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [bookSlug]
              properties:
                bookSlug: {type: string}
      responses:
        '201':
          description: Orden pendiente con URL de Checkout Pro.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Order'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '422':
          description: El libro no está publicado o no tiene precio válido.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '503':
          description: Mercado Pago no está configurado.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrder
      security: [{sessionCookie: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Orden perteneciente al usuario autenticado.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Order'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '404':
          description: Orden inexistente o perteneciente a otro usuario.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/webhooks/mercadopago:
    post:
      tags: [Webhooks]
      operationId: processMercadoPagoWebhook
      description: Valida x-signature y reconsulta el pago; no confía en el cuerpo recibido.
      parameters:
        - {name: data.id, in: query, required: true, schema: {type: string}}
        - {name: type, in: query, schema: {type: string, const: payment}}
        - {name: X-Signature, in: header, required: true, schema: {type: string}}
        - {name: X-Request-ID, in: header, required: true, schema: {type: string}}
      responses:
        '204': {description: Notificación procesada o pago ajeno ignorado de forma segura.}
        '401':
          description: Firma ausente, inválida o vencida.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '502':
          description: Fallo transitorio al consultar o persistir el pago.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  /api/v1/admin/books:
    get:
      tags: [Admin books]
      operationId: listAdminBooks
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Todos los libros, incluidos borradores y archivados.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BookList'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
    post:
      tags: [Admin books]
      operationId: createBook
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BookInput'}
      responses:
        '201':
          description: Libro creado.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
        '409': {$ref: '#/components/responses/BookSlugConflict'}
        '422': {$ref: '#/components/responses/BookValidationFailed'}
  /api/v1/admin/books/{identifier}:
    parameters:
      - $ref: '#/components/parameters/BookIdentifier'
    get:
      tags: [Admin books]
      operationId: getAdminBook
      security: [{sessionCookie: []}]
      responses:
        '200':
          description: Libro por UUID o slug.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
        '404': {$ref: '#/components/responses/BookNotFound'}
    put:
      tags: [Admin books]
      operationId: updateBook
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BookInput'}
      responses:
        '200':
          description: Libro actualizado.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
        '404': {$ref: '#/components/responses/BookNotFound'}
        '409': {$ref: '#/components/responses/BookSlugConflict'}
        '422': {$ref: '#/components/responses/BookValidationFailed'}
    delete:
      tags: [Admin books]
      operationId: archiveBook
      description: Cambia el estado a ARCHIVED; no elimina registros físicamente.
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
      responses:
        '204': {description: Libro archivado.}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
        '404': {$ref: '#/components/responses/BookNotFound'}
  /api/v1/admin/books/{identifier}/ebook:
    parameters:
      - $ref: '#/components/parameters/BookIdentifier'
    put:
      tags: [Admin books]
      operationId: uploadBookEbook
      description: Reemplaza el archivo privado del libro después de validar extensión, MIME y contenido.
      security: [{sessionCookie: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        '200':
          description: Archivo almacenado con una clave física opaca.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EbookUpload'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/AdminRequired'}
        '404': {$ref: '#/components/responses/BookNotFound'}
        '413':
          description: El archivo supera el límite configurado.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
        '422':
          description: El archivo no es un PDF o EPUB válido.
          content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nmp_session
  parameters:
    BookSlug:
      name: slug
      in: path
      required: true
      schema: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
    BookIdentifier:
      name: identifier
      in: path
      required: true
      description: UUID o slug del libro.
      schema: {type: string}
    Origin:
      name: Origin
      in: header
      required: true
      schema: {type: string, format: uri}
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: {type: string, const: ok}
    BookStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    User:
      type: object
      required: [id, email, displayName, pictureUrl, isAdmin]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        displayName: {type: string}
        pictureUrl: {type: string}
        isAdmin: {type: boolean}
    OrderStatus:
      type: string
      enum: [PENDING, PAID, CANCELLED, EXPIRED]
    OrderItem:
      type: object
      required: [bookId, bookSlug, bookTitle, unitPriceMinorUnits, quantity, currency]
      properties:
        bookId: {type: string, format: uuid}
        bookSlug: {type: string}
        bookTitle: {type: string}
        unitPriceMinorUnits: {type: integer, format: int64, minimum: 0}
        quantity: {type: integer, minimum: 1}
        currency: {type: string, minLength: 3, maxLength: 3}
    Order:
      type: object
      required: [id, status, totalMinorUnits, currency, items, createdAt, updatedAt, paidAt]
      properties:
        id: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/OrderStatus'}
        totalMinorUnits: {type: integer, format: int64, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3}
        checkoutUrl: {type: string, format: uri}
        items:
          type: array
          items: {$ref: '#/components/schemas/OrderItem'}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        paidAt: {type: [string, 'null'], format: date-time}
    BookInput:
      type: object
      additionalProperties: false
      required: [slug, title, priceMinorUnits]
      properties:
        slug: {type: string, maxLength: 160, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
        title: {type: string, maxLength: 200}
        subtitle: {type: string}
        authorName: {type: string}
        category: {type: string}
        variant: {type: string, enum: [gold, blue], default: blue}
        shortDescription: {type: string, maxLength: 1000}
        priceMinorUnits: {type: integer, format: int64, minimum: 0}
        currency: {type: string, minLength: 3, maxLength: 3, default: ARS}
        isbn: {type: string}
        publicationDate: {type: [string, 'null'], format: date}
        publicationDateLabel: {type: string}
        format: {type: string}
        fileSizeBytes: {type: [integer, 'null'], format: int64, minimum: 0}
        coverMediaId: {type: [string, 'null'], format: uuid}
        coverCaption: {type: string}
        status: {$ref: '#/components/schemas/BookStatus'}
        seoTitle: {type: string}
        seoDescription: {type: string}
        seoIndexable: {type: boolean, default: true}
    Book:
      allOf:
        - $ref: '#/components/schemas/BookInput'
        - type: object
          required: [id, status, hasCover, createdAt, updatedAt]
          properties:
            id: {type: string, format: uuid}
            hasCover: {type: boolean}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
            publishedAt: {type: [string, 'null'], format: date-time}
    BookList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Book'}
        total: {type: integer, minimum: 0}
    LibraryBook:
      type: object
      required: [id, slug, title, format, purchasedAt, downloadAvailable]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        title: {type: string}
        coverMediaId: {type: [string, 'null'], format: uuid}
        format: {type: string}
        fileSizeBytes: {type: [integer, 'null'], format: int64, minimum: 0}
        purchasedAt: {type: string, format: date-time}
        downloadAvailable: {type: boolean}
    LibraryBookList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/LibraryBook'}
        total: {type: integer, minimum: 0}
    EbookUpload:
      type: object
      required: [bookId, filename, mediaType, sizeBytes]
      properties:
        bookId: {type: string, format: uuid}
        filename: {type: string}
        mediaType: {type: string, enum: [application/pdf, application/epub+zip]}
        sizeBytes: {type: integer, format: int64, minimum: 1}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: {type: string}
            message: {type: string}
            details: {}
  responses:
    DatabaseUnavailable:
      description: PostgreSQL no está disponible.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    AuthenticationRequired:
      description: Falta una sesión válida.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    AdminRequired:
      description: La sesión no corresponde al administrador.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    BookNotFound:
      description: El libro no existe o no está publicado en una ruta pública.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    BookSlugConflict:
      description: El slug ya pertenece a otro libro.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    BookValidationFailed:
      description: Los datos del libro no cumplen el contrato.
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}